backend/requirements.txt
fastapi
uvicorn[standard]
pdf2image
pillow
pytesseract
python-multipart
requests
pydantic
python-dotenv
openai
fastapi
uvicorn[standard]
pdf2image
pillow
pytesseract
python-multipart
requests
pydantic
python-dotenv
openai
azure-ai-openai
streamlit
pdf2image
pillow
requests
python-docx